Fix rendering of 0.5 period

+++ test/test_regression.py
@@ -10,7 +10,7 @@ def make_regression():
         w.lane.period = test.period
         w.lane.hscale = test.hscale
         w.lane.phase = test.phase
-        output = w.parse_wave_lane(test.wave, int(test.period * test.hscale - 1))
+        output = w.parse_wave_lane(test.wave, test.period * test.hscale - 1)
         assert(output == test.expected)
